deploy 11/4

+++ app/GraphQL/Mutations/CartMutations.php
@@ -26,6 +26,7 @@ class CartMutations
         ];
         $args['user_id'] = Auth::id();
         $products = $args['products'];
+        $cart = Cart::create($args);
         foreach ($products as $product){
             $project = DevProject::find($product['id']);
             $project->purchases += 1 ;
@@ -41,11 +42,12 @@ class CartMutations
                 'user_sell' => $project->user->id,
                 'user_buy' => Auth::id(),
                 'status' => 'Chưa tiếp nhận',
+                'cart_id' => $cart->id,
                 'split_id' => $split->id,
             ]);
         }
         SendEmail::dispatch($message, [Auth::user()]);
-        return Cart::create($args);
+        return $cart;
     }
     function editCart($_, array $args){
         if(!isset($args['id'])) throw(new Exception('missing id value'));
@@ -54,6 +56,34 @@ class CartMutations
             ->update($args);
     }
     function deleteCart($_, array $args){
+        if(isset($args['force_delete']) && $args['force_delete']) 
+            return Cart::destroy($args['id']) > 0 ? true : false;
+        $cart = Cart::find($args['id']);
+        if($cart->user->id != Auth::id()){
+            $error = \Illuminate\Validation\ValidationException::withMessages([
+                'input.id' => ["Permission denied."],
+            ]);
+            throw $error;
+        }
+        $args = array_diff_key($args, array_flip(['directive']));
         return Cart::destroy($args['id']) > 0 ? true : false;
+        
+    }
+    function updateStatus($_, array $args){
+        $args = $args['input'];
+        try {
+            if(isset($args['force_update']) && $args['force_update']){
+                $buy_sell = ProjectSellBuy::where('cart_id', $args['cart_id']);
+            }else{
+                $buy_sell = ProjectSellBuy::where('project_id', $args['project_id'])
+                    ->where('cart_id', $args['cart_id'])->limit(1);
+            }
+
+        
+        return $buy_sell->update(['status' => $args['status']]);
+        } catch (\Exception $e) {
+            throw new Exception($e->getMessage());
+        }
+        
     }
 }
