+++ app/GraphQL/Mutations/DevProjectMutations.php
@@ -38,6 +38,15 @@ class DevProjectMutations
     }
     public function deleteDevProject($_, array $args): bool
     {
+        if(isset($args['force_delete']) && $args['force_delete']) 
+            return DevProject::destroy($args['id']) > 0 ? true : false;
+        $dev = DevProject::find($args['id']);
+        if($dev->user->id != Auth::id()){
+            $error = \Illuminate\Validation\ValidationException::withMessages([
+                'input.id' => ["Permission denied."],
+            ]);
+            throw $error;
+        }
         return DevProject::destroy($args['id']) > 0 ? true : false;
     }
     public function upsertDevProject($_, array $args){
